+++ src/index.ts
@@ -6,13 +6,12 @@ export * from './criteria';
 export * from './decorators';
 export * from './exceptions';
 export * from './mapping';
-export * from './mysqlStore';
 export * from './okResult';
 export * from './store';
 export * from './types';
 
-import * as mysql from './dialects/mysql';
-export const MySQL = mysql;
+import * as _mysql from './dialects/mysql';
+export const mysql = _mysql;
 
 import * as symbols from './symbols';
 export const Symbols = symbols;