+++ src/mysql.ts
@@ -1,2 +1,1 @@
 export * from './dialects/mysql';
-export * from './mysqlStore';